+++ back_end/src/controllers/event.controller.ts
@@ -28,11 +28,22 @@ export const findOne: RequestHandler = async (req, res) => {
  * POST /events
  */
 export const create: RequestHandler = async (req, res) => {
-  const { title, description, dateEvent, authorId } = req.body;
-  const newEvent = await prisma.event.create({
-    data: { title, description, dateEvent, authorId }
-  });
-  res.status(201).json(newEvent);
+  const { title, description, dateEvent } = req.body;
+	const { sub, roles } = (req as any).user;
+
+	let finalAuthorId = req.body.authorId;
+	if (roles.includes('author')) {
+		finalAuthorId = sub;
+	}
+	try {
+		const newEvent = await prisma.event.create({
+			data: { title, description, dateEvent, authorId: finalAuthorId}
+		})
+		res.status(201).json(newEvent);
+	} catch (err: any) {
+		console.error('Erreur création événement :', err);
+		res.status(500).json({ error: 'Impossible de créer un événement' });
+	}
 };
 
 /**
@@ -40,19 +51,47 @@ export const create: RequestHandler = async (req, res) => {
  */
 export const update: RequestHandler = async (req, res) => {
   const { id } = req.params;
-  const { title, description, dateEvent, authorId } = req.body;
-  const updated = await prisma.event.update({
-    where: { id },
-    data: { title, description, dateEvent, authorId }
-  });
-  res.json(updated);
+	const { title, description, dateEvent } = req.body;
+	const { sub, roles } = (req as any).user;
+
+	let finalAuthorId = req.body.authorId;
+	if (roles.includes('author')) {
+		finalAuthorId = sub;
+	}
+
+	try {
+		const updated = await prisma.event.update({
+			where: {id},
+			data: { title, description, dateEvent, authorId: finalAuthorId }
+		});
+		res.json(updated);
+	} catch (err: any) {
+		console.error('Erreur mise à jour événement :', err);
+		res.status(500).json({ error: "Impossible de mettre à jour l'événement" });
+	}
 };
 
 /**
- * DELETE /events/:id
+ * DELETE /events/:id uniquement l'auteur
  */
 export const remove: RequestHandler = async (req, res) => {
   const { id } = req.params;
-  await prisma.event.delete({ where: { id } });
-  res.sendStatus(204);
+	const { sub, roles } = (req as any).user;
+
+	try {
+		const event = await prisma.event.findUnique({ where: { id } });
+		if (!event) {
+			res.status(404).json({ error: 'Événement non trouvé' });
+			return;
+		}
+		if (roles.includes('author') && event.authorId !== sub) {
+			res.status(403).json({ error: 'Impossible de supprimer cet événement' });
+			return;
+		}
+		await prisma.event.delete({ where: { id } });
+		res.sendStatus(204);
+	} catch (err: any) {
+		console.error('Erreur suppression événement :', err);
+		res.status(500).json({ error: "Impossible de supprimer l'événement" })
+	}
 };
